+++ example/cli/src/app/fulltree/fulltree.component.ts
@@ -12,8 +12,8 @@ const actionMapping:IActionMapping = {
     },
     click: (tree, node, $event) => {
       $event.shiftKey
-        ? TREE_ACTIONS.TOGGLE_SELECTED_MULTI(tree, node, $event)
-        : TREE_ACTIONS.TOGGLE_SELECTED(tree, node, $event)
+        ? TREE_ACTIONS.TOGGLE_ACTIVE_MULTI(tree, node, $event)
+        : TREE_ACTIONS.TOGGLE_ACTIVE(tree, node, $event)
     }
   },
   keys: {