+++ lib/components/tree-node-checkbox.component.ts
@@ -1,0 +1,21 @@
+import { Component, Input, ViewEncapsulation } from '@angular/core';
+import { TreeNode } from '../models/tree-node.model';
+
+@Component({
+  selector: 'tree-node-checkbox',
+  encapsulation: ViewEncapsulation.None,
+  styles: [],
+  template: `
+    <ng-container *mobxAutorun="{dontDetach: true}">
+      <input
+        class="tree-node-checkbox"
+        type="checkbox"
+        (click)="node.mouseAction('checkboxClick', $event)"
+        [checked]="node.isSelected"
+        [indeterminate]="node.isPartiallySelected"/>
+    </ng-container>
+  `
+})
+export class TreeNodeCheckboxComponent {
+  @Input() node: TreeNode;
+}