+++ lib/components/tree-node-wrapper.component.ts
@@ -7,6 +7,7 @@ import { TreeNode } from '../models/tree-node.model';
   styles: [],
   template: `
       <div *ngIf="!templates.treeNodeWrapperTemplate" class="node-wrapper" [style.padding-left]="node.getNodePadding()">
+        <tree-node-checkbox *ngIf="node.options.useCheckbox" [node]="node"></tree-node-checkbox>
         <tree-node-expander [node]="node"></tree-node-expander>
         <div class="node-content-wrapper"
           [class.node-content-wrapper-active]="node.isActive"