+++ lib/models/tree-options.model.ts
@@ -10,8 +10,9 @@ export interface IActionHandler {
 }
 
 export const TREE_ACTIONS = {
-  TOGGLE_SELECTED: (tree: TreeModel, node: TreeNode, $event: any) => node && node.toggleActivated(),
-  TOGGLE_SELECTED_MULTI: (tree: TreeModel, node: TreeNode, $event: any) => node && node.toggleActivated(true),
+  TOGGLE_ACTIVE: (tree: TreeModel, node: TreeNode, $event: any) => node && node.toggleActivated(),
+  TOGGLE_ACTIVE_MULTI: (tree: TreeModel, node: TreeNode, $event: any) => node && node.toggleActivated(true),
+  TOGGLE_SELECTED: (tree: TreeModel, node: TreeNode, $event: any) => node && node.toggleSelected(),
   SELECT: (tree: TreeModel, node: TreeNode, $event: any) => node.setIsActive(true),
   DESELECT: (tree: TreeModel, node: TreeNode, $event: any) => node.setIsActive(false),
   FOCUS: (tree: TreeModel, node: TreeNode, $event: any) => node.focus(),
@@ -34,10 +35,11 @@ export const TREE_ACTIONS = {
 
 const defaultActionMapping: IActionMapping = {
   mouse: {
-    click: TREE_ACTIONS.TOGGLE_SELECTED,
+    click: TREE_ACTIONS.TOGGLE_ACTIVE,
     dblClick: null,
     contextMenu: null,
     expanderClick: TREE_ACTIONS.TOGGLE_EXPANDED,
+    checkboxClick: TREE_ACTIONS.TOGGLE_SELECTED,
     drop: TREE_ACTIONS.MOVE_NODE
   },
   keys: {
@@ -45,8 +47,8 @@ const defaultActionMapping: IActionMapping = {
     [KEYS.LEFT]: TREE_ACTIONS.DRILL_UP,
     [KEYS.DOWN]: TREE_ACTIONS.NEXT_NODE,
     [KEYS.UP]: TREE_ACTIONS.PREVIOUS_NODE,
-    [KEYS.SPACE]: TREE_ACTIONS.TOGGLE_SELECTED,
-    [KEYS.ENTER]: TREE_ACTIONS.TOGGLE_SELECTED
+    [KEYS.SPACE]: TREE_ACTIONS.TOGGLE_ACTIVE,
+    [KEYS.ENTER]: TREE_ACTIONS.TOGGLE_ACTIVE
   }
 };
 
@@ -56,6 +58,7 @@ export interface IActionMapping {
     dblClick?: IActionHandler,
     contextMenu?: IActionHandler,
     expanderClick?: IActionHandler,
+    checkboxClick?: IActionHandler,
     dragStart?: IActionHandler,
     drag?: IActionHandler,
     dragEnd?: IActionHandler,
@@ -84,6 +87,7 @@ export class TreeOptions {
   get scrollOnSelect(): boolean { return this.options.scrollOnSelect === undefined ? true : this.options.scrollOnSelect; }
   get rtl(): boolean { return !!this.options.rtl; }
   get rootId(): any {return this.options.rootId; }
+  get useCheckbox(): boolean { return this.options.useCheckbox; }
   actionMapping: IActionMapping;
 
   constructor(private options: ITreeOptions = {}) {